refactor: consolidate economy and ecology data
