+++ assets/economy/currency.js
