+++ assets/economy/items.json
@@ -20,7 +20,7 @@
     "lc_eff": 0.9,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.2,
-    "net_profit_cp": 1.4,
+    "net_profit_cp": 1.468,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -63,7 +63,7 @@
     "lc_eff": 1.575,
     "effective_tax_pct": 0.06,
     "tax_amount_cp": 0.3,
-    "net_profit_cp": 1.7,
+    "net_profit_cp": 2.115,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -192,7 +192,7 @@
     "lc_eff": 1.62,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.3,
-    "net_profit_cp": 1.4,
+    "net_profit_cp": 1.844,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -235,7 +235,7 @@
     "lc_eff": 2.7,
     "effective_tax_pct": 0.06,
     "tax_amount_cp": 0.45,
-    "net_profit_cp": 1.65,
+    "net_profit_cp": 2.37,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -880,7 +880,7 @@
     "lc_eff": 3.78,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.6,
-    "net_profit_cp": 2.27,
+    "net_profit_cp": 2.28,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -11791,7 +11791,7 @@
     "lc_eff": 1.8,
     "effective_tax_pct": 0.04,
     "tax_amount_cp": 0.4,
-    "net_profit_cp": 2.1,
+    "net_profit_cp": 2.66,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -11836,7 +11836,7 @@
     "lc_eff": 2.7,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.6,
-    "net_profit_cp": 1.4,
+    "net_profit_cp": 2.42,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -11881,7 +11881,7 @@
     "lc_eff": 2.7,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.6,
-    "net_profit_cp": 1.4,
+    "net_profit_cp": 2.42,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
@@ -12016,7 +12016,7 @@
     "lc_eff": 2.7,
     "effective_tax_pct": 0.05,
     "tax_amount_cp": 0.6,
-    "net_profit_cp": 1.4,
+    "net_profit_cp": 2.42,
     "duty": false,
     "duty_pct": 0,
     "duty_free": false,
