--- assets/economy/util.js
@@ -1,4 +1,0 @@
-export function clamp(v, min, max){
-  return Math.min(Math.max(v, min), max);
-}
-
