--- data/animalVariants.json
@@ -1,40 +1,0 @@
-[
-  {
-    "animal_id": "chicken",
-    "label": "Common",
-    "notes": ""
-  },
-  {
-    "animal_id": "chicken",
-    "label": "Fine",
-    "notes": ""
-  },
-  {
-    "animal_id": "chicken",
-    "label": "Trained",
-    "notes": "Accustomed to handling"
-  },
-  {
-    "animal_id": "chicken",
-    "label": "Champion",
-    "notes": "Exceptional breeder/show"
-  },
-  {
-    "animal_id": "draft-horse",
-    "label": "Common"
-  },
-  {
-    "animal_id": "draft-horse",
-    "label": "Fine"
-  },
-  {
-    "animal_id": "draft-horse",
-    "label": "Trained",
-    "notes": "Harness trained"
-  },
-  {
-    "animal_id": "draft-horse",
-    "label": "Champion",
-    "notes": "Superior strength/endurance"
-  }
-]
