--- data/plantSKUs.json
@@ -1,354 +1,0 @@
-[
-  {
-    "plant_id": "apple-tree",
-    "sku": "Apples",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "peppercorn",
-    "sku": "Peppercorns",
-    "units": "sack",
-    "luxury_tiers": [
-      "Common",
-      "Fine",
-      "Luxury",
-      "Arcane"
-    ]
-  },
-  {
-    "plant_id": "saffron",
-    "sku": "Saffron",
-    "units": "dram",
-    "luxury_tiers": [
-      "Common",
-      "Fine",
-      "Luxury",
-      "Arcane"
-    ]
-  },
-  {
-    "plant_id": "chanterelle",
-    "sku": "Chanterelles",
-    "units": "basket",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "cabbage",
-    "sku": "Cabbage",
-    "units": "head",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "apple-tree",
-    "sku": "Red Apples",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "apple-tree",
-    "sku": "Green Apples",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "apple-tree",
-    "sku": "Golden Apples",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "potato",
-    "sku": "Russet Potatoes",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "potato",
-    "sku": "Red Potatoes",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "potato",
-    "sku": "Yukon Gold Potatoes",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "onion",
-    "sku": "Yellow Onions",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "onion",
-    "sku": "Red Onions",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "onion",
-    "sku": "White Onions",
-    "units": "sack",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "broad-beans",
-    "sku": "Pinto Beans",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "broad-beans",
-    "sku": "Black Beans",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "chickpeas",
-    "sku": "Chickpeas",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "lentils",
-    "sku": "Brown Lentils",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "lentils",
-    "sku": "Red Lentils",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "lentils",
-    "sku": "Green Lentils",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "cabbage",
-    "sku": "Green Cabbage",
-    "units": "each",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "cabbage",
-    "sku": "Red Cabbage",
-    "units": "each",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "orange",
-    "sku": "Navel Oranges",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "orange",
-    "sku": "Blood Oranges",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "orange",
-    "sku": "Seville Oranges",
-    "units": "dozen",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "grapes",
-    "sku": "Red Grapes",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "grapes",
-    "sku": "White Grapes",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "carrot",
-    "sku": "Orange Carrots",
-    "units": "bunch",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "tomato",
-    "sku": "Red Tomatoes",
-    "units": "lb",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "cucumber",
-    "sku": "Green Cucumbers",
-    "units": "each",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "lettuce",
-    "sku": "Green Lettuce",
-    "units": "each",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  },
-  {
-    "plant_id": "strawberries",
-    "sku": "Red Strawberries",
-    "units": "pint",
-    "tiers": [
-      "Low Inn",
-      "Common",
-      "Fine",
-      "High Table"
-    ]
-  }
-]
