--- reports/import_economy_catalog_2025-09-11T20-12-18-253Z.json
@@ -1,6 +1,0 @@
-{
-  "inserted": 132,
-  "updated": 0,
-  "skipped": 212,
-  "errors": []
-}
