+++ tests/economy_import/net_profit.test.js
@@ -1,6 +1,6 @@
 import test from 'node:test';
 import assert from 'node:assert';
-import data from '../../assets/data/economy_items.json' assert { type: 'json' };
+import data from '../../assets/economy/items.json' assert { type: 'json' };
 
 test('net profit matches cost formula', () => {
   for (const item of data) {
